Refactor LLM bullshit

+++ src/utils/macos.rs
@@ -6,7 +6,7 @@ macro_rules! schedule_timer {
   ($interval:expr, $target:expr, $selector:ident) => {{
     let timer = unsafe {
       objc2_foundation::NSTimer::timerWithTimeInterval_target_selector_userInfo_repeats(
-        $interval, $target, objc2::sel!($selector:), None, true,
+        $interval as f64, $target, objc2::sel!($selector:), None, true,
       )
     };
